separated allQandA's component from main QandA component.
