+++ client/src/index.js
@@ -3,7 +3,6 @@
 import ReactDOM from 'react-dom';
 import React from 'react';
 import App from './components/App.jsx';
-import QandA from './components/Q&A.jsx';
 
 ReactDOM.render(
   <App />,
