+++ webpack.config.js
@@ -10,13 +10,12 @@ module.exports = {
     path: DIST_DIR,
   },
   module: {
-
     rules: [
       {
         test: /\.js(x?)$/,
         exclude: /node_modules/,
         use: {
-          loader: ['babel-loader', 'style-loader!css-loader'],
+          loader: 'babel-loader',
           options: {
             presets: ['@babel/preset-env', '@babel/preset-react'],
           }
@@ -25,3 +24,6 @@ module.exports = {
     ]
   }
 };
+
+
+
